remake pmbuniyap project done

+++ app/controllers/Admin.php
@@ -60,11 +60,15 @@ class Admin extends Controller
 
     public function tambah()
     {
-        // var_dump($_POST);
+    
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
         
         $data['cama'] = $_POST;
         $data['dokumen'] = $_FILES;
-        // $this->model('Admin_model')->addMahasiswa($data);
 
         if ($this->model('Admin_model')->addMahasiswa($data) > 0) {
             Flasher::setFlash('Berhasil', 'Menambah Data Mahasiswa !', 'success', 'admin');
@@ -80,6 +84,12 @@ class Admin extends Controller
 
     public function hapus($nisn = [])
     {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
         if ($this->model('Admin_model')->hapus($nisn) == 1) 
         {
             Flasher::setFlash('Berhasil', 'Menghapus Data Mahasiswa !', 'success', 'admin');
@@ -94,6 +104,14 @@ class Admin extends Controller
 
     public function getUbah()
     {
+
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+
         $data = $this->model('Admin_model')->getUbah($_POST['nisn']);
 
         echo json_encode($data);
@@ -101,6 +119,13 @@ class Admin extends Controller
 
     public function ubah()
     {
+
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
         if ($this->model('Admin_model')->ubah([$_POST, $_FILES]) > 0) {
             Flasher::setFlash('Berhasil', 'Mengubah Data Mahasiswa !', 'success', 'admin');
             header('Location: '. BASEURL .'/admin');
@@ -114,10 +139,15 @@ class Admin extends Controller
 
 
     // halaman prodi admin
-
-
     public function prodi()
     {
+
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
         $data = $this->model('Cama_model')->getProdi();
 
         $this->view('admin/dashboard/templates/header');
@@ -127,6 +157,13 @@ class Admin extends Controller
 
     public function addProdi()
     {
+
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
         if ($this->model('Admin_prodi_model')->addProdi($_POST) > 0) {
             Flasher::setFlash('Berhasil', 'Menambah Program Studi !', 'success', 'admin/prodi');
             header('Location: '. BASEURL .'/admin/prodi');
@@ -140,6 +177,13 @@ class Admin extends Controller
 
     public function deleteProdi($kode_prodi)
     {
+
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
         if ($this->model('Admin_prodi_model')->deleteProdi($kode_prodi) > 0) {
             Flasher::setFlash('Berhasil', 'Menghapus Program Studi !', 'success', 'admin/prodi');
             header('Location: '. BASEURL .'/admin/prodi');
@@ -153,19 +197,174 @@ class Admin extends Controller
 
     public function getEditProdi()
     {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
         echo json_encode($this->model('Admin_prodi_model')->getEditProdi($_POST['kode']));
     }
 
     public function editProdi()
     {
-        if ($this->model('Admin_prodi_model')->editProdi($_POST)) {
+
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+        if ($this->model('Admin_prodi_model')->editProdi($_POST) > 0) {
             Flasher::setFlash('Berhasil', 'Edit Program Studi !', 'success', 'admin/prodi');
             header('Location: '. BASEURL .'/admin/prodi');
             exit;
         }else {
-            Flasher::setFlash('Gagal', 'Edit Program Studi !', 'danger', 'admin/prodi');
+            Flasher::setFlash('Tidak Ada', 'Data Yang Berubah !', 'warning', 'admin/prodi');
             header('Location: '. BASEURL .'/admin/prodi');
             exit;
         }
+    }
+
+    //halaman jadwal admin
+
+
+    public function jadwal()
+    {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+        $data = $this->model('Admin_jadwal_model')->getJadwal();
+
+        $this->view('admin/dashboard/templates/header');
+        $this->view('admin/dashboard/jadwal', $data);
+        $this->view('admin/dashboard/templates/footer');
+    }
+
+    public function addJadwal()
+    {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+        if ($this->model('Admin_jadwal_model')->addJadwal($_POST)) {
+            Flasher::setFlash('Berhasil', 'Menambahkan Jadwal !', 'success', 'admin/jadwal');
+            header('Location: '. BASEURL .'/admin/jadwal');
+            exit;
+        }else {
+            Flasher::setFlash('Gagal', 'Menambahkan Jadwal !', 'danger', 'admin/jadwal');
+            header('Location: '. BASEURL .'/admin/jadwal');
+            exit;
+        }
+    }
+
+    public function deleteJadwal($id)
+    {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+        if ($this->model('Admin_jadwal_model')->deleteJadwal($id)) {
+            Flasher::setFlash('Berhasil', 'Menghapus Jadwal !', 'success', 'admin/jadwal');
+            header('Location: '. BASEURL .'/admin/jadwal');
+            exit;
+        }else {
+            Flasher::setFlash('Gagal', 'Menghapus Jadwal !', 'danger', 'admin/jadwal');
+            header('Location: '. BASEURL .'/admin/jadwal');
+            exit;
+        }
+    }
+
+    public function getUbahJadwal()
+    {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+        echo json_encode($this->model('Admin_jadwal_model')->getUbahJadwal($_POST['id']));
+    }
+
+    public function editJadwal()
+    {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+        if ($this->model('Admin_jadwal_model')->editJadwal($_POST) > 0) {
+            Flasher::setFlash('Berhasil', 'Mengubah Jadwal !', 'success', 'admin/jadwal');
+            header('Location: '. BASEURL .'/admin/jadwal');
+            exit;
+        }else {
+            Flasher::setFlash('Tidak Ada', 'Data Yang Berubah !', 'warning', 'admin/jadwal');
+            header('Location: '. BASEURL .'/admin/jadwal');
+            exit;
+        }
+    }
+
+
+    // dokumen
+
+    public function dokumen()
+    {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+      
+        $data = $this->model('Admin_model')->getAllDokumen();
+
+        $this->view('admin/dashboard/templates/header');
+        $this->view('admin/dashboard/dokumen', $data);
+        $this->view('admin/dashboard/templates/footer');
+    }
+
+
+
+    // seleksi
+
+
+    public function seleksi()
+    {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+
+        $data = $this->model('Admin_seleksi_model')->getData();
+
+        $this->view('admin/dashboard/templates/header');
+        $this->view('admin/dashboard/seleksi', $data);
+        $this->view('admin/dashboard/templates/footer');
+    }
+
+
+    public function getStatus()
+    {
+        if (!(isset($_SESSION['auth']))) {
+            unset($_SESSION['auth']);
+            header('Location:'. BASEURL .'/');
+            exit;
+        }
+        
+        if ($this->model('Admin_seleksi_model')->getStatus($_POST))
+        {
+            header('Location:'. BASEURL .'/admin/seleksi');
+            exit;
+        }
+
     }
 }
